requirements.txt
appdirs==1.4.3
astroid==2.3.3
attrs==19.3.0
bcrypt==3.1.7
black==19.10b0
casbin==0.8.4
casbin-peewee-adapter==0.0.2
cffi==1.14.0
Click==7.0
Flask==1.1.1
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
pathspec==0.7.0
peewee==3.13.1
pycparser==2.19
PyJWT==1.7.1
pylint==2.4.4
regex==2020.1.8
simpleeval==0.9.10
six==1.14.0
toml==0.10.0
typed-ast==1.4.1
Werkzeug==1.0.0
wrapt==1.11.2
